backend/requirements.txt
django
djangorestframework
environs
Pillow
django-cors-headers